Removed Arachnids from having Collective Minds, and reverted change that caused UI Issue.
Commit 068c1d7546070f9a80aae8e620bc36088a0c9422 removed code that displayed the name of the available collective minds correctly. It would use placeholder text instead the localized name.

Additionally, a bug I have found is that when you become infected as a Zombie, the C menu does not update to say you have access to the Zombie collective mind. It will only appear once you attempt to use the Zombie channel. Then it will appear in the C menu correctly.