Changed required Diona Tag, commented out Diona Hivemind
The original Diona tag was simply "Diona". While it included a comment to tell the reader it belonged to the Hivemind code, it has been changed to "DionaHivemind" to make it more obvious.

Instead of commenting out the DionaHivemind tag, the Diona hivemind in collective_minds.yml has been commented out instead. Meaning there is no longer a Diona hivemind, even if someone has the required tag.