Card System
This is the start of the system.
The only functionality is that when you press E on a card, it flips the card.
I tried to get it to work when pressing F instead of E. But shit was too complicated and Roomba would have exploded at the thought of me touching atmos code.