Fix layering issue (controversial)
I'm editing the sprites to make to remove far away leg/arm from the side sprites.

Meaning if a Resomi is looking to the left of the screen, the far away leg (right leg) sprite has pretty much been erased. I thinnk this fixes one of the issues with layering.